django-api/requirements.txt
Django==3.2.13
djangorestframework==3.13.1
PyJWT==2.4.0
django-cors-headers==3.13.0
gunicorn==20.1.0
django-environ==0.8.1
pymongo==4.5.0
coinbase-commerce==1.0.1
python-dotenv==1.0.0